fix: enable Android predictive back gesture support
Added android:enableOnBackInvokedCallback="true" to AndroidManifest.xml to properly handle the back gesture navigation on Android 13+ devices and remove the OnBackInvokedCallback warning.

The caching mechanism added for registering the UI state for rebuil function using mixin/UIStateAware and its manager. This is done to avoid Corrupted UI elements when app resumes after existing it for any reason.
Also this state can be a helper for keeping the latest progress user had in the app.